Implement adjustable test threshold
Adds a new user settable parameter `THRESHOLD` which determines how strict the comparison has to be to pass an image comparison test.  `THRESHOLD` takes on values of 0 to 1 with a 1 being a perfect pixel-by-pixel match.
To facilitate this new setting, test of comparison result `metric` vs. `THRESHOLD` is now done using `bc` for floating point precision.  This replaces previous coercion of value to integer (round to 3 decimal places and multiply by 1000) which enabled `bash`'s internal integer comparison operators to do the test.
The default value of 0.9985 is designed to imitate the previous test limit of 999 (it's the smallest value that would be coerced to 999).