Differentiation error message
When an expectation image cannot be found the image comparison is skipped and the test fails.  Previously, however, this was reported as a "Results differ" failure, implying that the comparison was run.  This change makes it so that the reported error message indicates the real cause of the test failure (missing expectation image).