Remove empty image-cache directory
When the `convert` command to populate an image-cache directory fails, we remove the empty directory.  We do this because subsequent runs of the script will use the existence (and timestamp) on the directory to determine if the image-cache can be used for a particular test and an empty directory would falsely indicate that the image-cache was available.