Update Alegreya font files
LuaTeX switched to HarfBuzz font rendering in TeXLive 2020 and this requires that the Alegreya test fonts be updated.  Without the update there are issues with the rendering of small caps and accented characters (particularly \ae and \oe).
This commit contains only the updated font files.  Subsequent commits will deal with the minor kerning issues that this change will inevitably create.