Refine kpse tests exit status
If the version of Gregorio being tested doesn't make use of the kpathsea libriaries, then these tests are invalid tests, so the tests themselves should note that with their exit state.
However, we don't want the tests to fail and clutter up the results, so we auto-pass them (just as before) but print a note about the autopass to the screen.