Exit state tests
While one could argue that the original tests were based really tests of the exit state, I decided not to rewrite the tests in a way that makes explicit that we are testing the exit state and what we expect the exit state to be.  Since we only use exit states of 0 (success) and 1 (failure) that may seem like overkill, but this leaves open the possibility of us making a more expansive use of exit states in the future.