Update kpsewhich tests
These tests require that the kpathsea libraries are *not* used, but rather kpsewhich is invoked from PATH.  If this condition is not met, these test will be automatically passed as invalid tests.
If the tests are valid, we check the stderr and exit to make sure they are correct.  We don't check the file output (there are plenty of other tests doing that), nor do we check stdout (there is none).  This is all in keeping with the way the other scripted tests have been redone.